Try making it an int instead of double (we really going dumb here) It makes no sense it works for one value, and not the other, it is IDENTICAL

+++ src/main/java/com/example/s3rekognition/controller/RekognitionController.java
@@ -234,7 +234,7 @@ public class RekognitionController implements ApplicationListener<ApplicationRea
 
     @Timed
     public void registerToMeter(String violationType, int violations, int nonViolations, String violationTypePercentage, int people) {
-        double violationsPercentage = ((double) violations / (violations + nonViolations)) * 100;
+        int violationsPercentage = (int)((double) violations / (violations + nonViolations)) * 100;
         meterRegistry.counter(violationType).increment(violations);
         meterRegistry.counter("violations_total").increment(violations);
         meterRegistry.gauge(violationTypePercentage, violationsPercentage);
